requirements/dev.txt
-r base.txt
bandit==1.9.2
boolean.py==5.0
CacheControl==0.14.4
colorama==0.4.6
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
django-stubs==5.2.8
django-stubs-ext==5.2.8
filelock==3.20.1
hypothesis==6.148.7
iniconfig==2.3.0
librt==0.7.4
license-expression==30.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
mypy==1.19.1
mypy_extensions==1.1.0
packageurl-python==0.17.6
packaging==25.0
pathspec==0.12.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.1
pluggy==1.6.0
py-serializable==2.1.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.2
pytest-django==4.11.1
PyYAML==6.0.3
rich==14.2.0
ruff==0.14.9
sortedcontainers==2.4.0
stevedore==5.6.0
tomli==2.3.0
tomli_w==1.2.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
Werkzeug==3.1.4
